# server/openapi/model_test_.go
/*
 * TraceTest
 *
 * OpenAPI definition for TraceTest endpoint and resources
 *
 * API version: 0.2.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Test struct {
	Id string `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Description string `json:"description,omitempty"`

	ServiceUnderTest TestServiceUnderTest `json:"serviceUnderTest,omitempty"`

	// Map using selector query as key, and an array of assertions as value
	Definition map[string][]Assertion `json:"definition,omitempty"`

	// Map using selector query as key, and an array of assertions as value
	ReferenceTestRunResult map[string][]Assertion `json:"referenceTestRunResult,omitempty"`
}

// AssertTestRequired checks if the required fields are not zero-ed
func AssertTestRequired(obj Test) error {
	if err := AssertTestServiceUnderTestRequired(obj.ServiceUnderTest); err != nil {
		return err
	}
	return nil
}

// AssertRecurseTestRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Test (e.g. [][]Test), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseTestRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aTest, ok := obj.(Test)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertTestRequired(aTest)
	})
}